Workaround bug in pcapng writer
The library uses a stupid counter instead of a map with interface
indexes. This leads to the faulty error:

Can't send statistics for non existent interface 14; have only 2 interfaces

This is caused by this line:

https://github.com/google/gopacket/blob/5e49d8961df06368385353c8be68fa1051e9fa95/pcapgo/ngwrite.go#L357

If you e.g. try to capture interface with interface index 14, then you
have to call the AddInterface() method 15 times. Instead the library
should use a map to track which interface indexes have already been
added to the pcap. For now we use the legacy format.